--- userData/IMemo.java
@@ -1,10 +1,0 @@
-package userData;
-
-import java.util.Date;
-
-public interface IMemo
-{
-    String getMemoText();
-
-    boolean dateCompare(Date date);
-}
