--- userData/IMemoManager.java
@@ -1,10 +1,0 @@
-package userData;
-
-public interface IMemoManager
-{
-    void isUserExists(Long userId);
-
-    void addMemoToUser(Long userId, Memo reminder);
-
-    void removeMemoFromUser(Long userId, Memo reminder);
-}
