--- userData/IUser.java
@@ -1,10 +1,0 @@
-package userData;
-
-public interface IUser
-{
-    void addMemo(Memo reminder);
-
-    void removeMemo(Memo reminder);
-
-    boolean isMemoExists();
-}
